Genai/langchain/history/requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
pypdf==3.17.1
python-dotenv==1.0.0
